move storage service and provide it in root; move app component to snake component
