+++ src/app/app.component.html
@@ -1,21 +1,1 @@
-<div class="game-container">
-  <div class="game-header">
-    <h3 class="logo">ng-Snake</h3>
-    <div class="score-block">
-      <h3 class="score" [ngClass]="{'new-best-score': newBestScore}">Score: {{ score }}</h3>
-      <h3 class="best-score" [ngClass]="{'new-best-score': newBestScore}">Best Score: {{ best_score }}</h3>
-    </div>
-  </div>
-  <div class="row" *ngFor="let column of board; let i = index;">
-    <div class="column" [ngStyle]="{'background-color': setCellColor(i, j)}"
-         *ngFor="let row of column; let j = index"></div>
-  </div>
-  <div class="start-button" [ngClass]="{'disable-clicks': gameStarted}" (click)="showMenu()">Start Game</div>
-  <div class="new-game-menu" *ngIf="showMenuChecker">
-    <span class="new-game-menu-label">Select Mode</span>
-    <div class="start-button new-game-button" (click)="newGame(mode.value)"
-         *ngFor="let mode of GameModesEnum | keyvalue">
-      {{ mode.value }}
-    </div>
-  </div>
-</div>
+<app-snake></app-snake>
