+++ src/app/common/index.ts
@@ -1,2 +1,4 @@
 export * from './enums/index';
 export * from './consts/index';
+export * from './services/index';
+
