+++ src/app/common/services/index.ts
@@ -1,0 +1,1 @@
+export * from './storage/storage.service'
