+++ src/app/snake/snake-board/snake-board.component.html
@@ -1,0 +1,1 @@
+<p>snake-board works!</p>
