+++ src/app/snake/snake-board/snake-board.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-snake-board',
+  templateUrl: './snake-board.component.html',
+  styleUrl: './snake-board.component.css'
+})
+export class SnakeBoardComponent {
+
+}
