+++ src/app/snake/snake-header/snake-header.component.html
@@ -1,0 +1,1 @@
+<p>snake-header works!</p>
