+++ src/app/snake/snake-header/snake-header.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-snake-header',
+  templateUrl: './snake-header.component.html',
+  styleUrl: './snake-header.component.css'
+})
+export class SnakeHeaderComponent {
+
+}
